+++ MKLoggingFiles/LogUnbufferedFileSink.h
@@ -16,7 +16,7 @@ public: // ILogSink
     void OutputString(const std::string& text) override;
 
 private:
-    void OutputData(const void* data, unsigned int size);
+    void OutputData(const void* data, size_t size);
     int m_FileDescriptor = 0;
 };
